More proger with the test
