+++ aiida/backends/tests/workflows.py
@@ -16,7 +16,7 @@ from aiida.workflows.test import WFTestEmpty
 from aiida.orm.implementation import get_workflow_info
 from aiida.workflows.test import WFTestSimpleWithSubWF
 from aiida.workflows.wf_demo import WorkflowDemo
-from aiida.daemon.workflowmanager import execute_steps
+from aiida.daemon.tasks import workflow_stepper
 
 
 class TestWorkflowBasic(AiidaTestCase):
@@ -154,29 +154,21 @@ class TestWorkflowBasic(AiidaTestCase):
         # it is a valid state
         self.assertIn(wf.get_state(), wf_states)
 
-    def test_old_wf_results(self):
-        class FailingWorkflowDemo(WorkflowDemo):
-            def generate_calc(self):
-                from aiida.orm import Code, Computer, CalculationFactory
-                from aiida.common.datastructures import calc_states
+    def test_failing_calc_in_wf(self):
+        from time import sleep
+        from aiida.daemon.workflowmanager import execute_steps
+        from aiida.workflows.test import WFTestSimple, WFTestSimpleWithSubWF
+        from aiida.orm.workflow import Workflow
+        from aiida.workflows.test import FailingWFTestSimple
 
-                CustomCalc = CalculationFactory(
-                    'simpleplugins.templatereplacer')
+        wf = FailingWFTestSimple()
+        wf.store()
 
-                computer = Computer.get("localhost")
-
-                calc = CustomCalc(computer=computer, withmpi=True)
-                calc.set_resources(
-                    {"num_machines": 1, "num_mpiprocs_per_machine": 1})
-                calc.store()
-                calc._set_state(calc_states.FAILED)
-
-                return calc
-
-        wf = FailingWorkflowDemo()
         wf.start()
         while wf.is_running():
+            # workflow_stepper()
             execute_steps()
+            sleep(1)
 
 
     def tearDown(self):
